refactor(runtime): extract native-streaming startup
Move the credential gate, the deferred librespot bring-up, and the
saved-device decision out of the tui-gated startup.rs into
runtime/streaming/, gated on streaming. The pure decision lives in
mod.rs, the bring-up in launch.rs.

Add a headless-streaming CI leg (telemetry,streaming; check + clippy)
so the startup builds with no terminal frontend.
